fix: mock HTTP in feedback test to prevent real GitHub issues
test_feedback_submit_success was hitting the real feedback API and
creating actual GitHub issues. Now mocks httpx like the other feedback
tests. Extracted formatting/parsing tests to test_formatting.py.
